+++ cvsgit/command/pull.py
@@ -25,6 +25,7 @@ class pull(Command):
             _("Verify the new HEAD revision and work tree against "
               "a fresh CVS checkout (does not work in a bare "
               "repository.)"))
+        self.add_no_skip_latest_option()
         self.add_authors_option()
         self.add_stop_on_unknown_author_option()
 
@@ -39,6 +40,7 @@ class pull(Command):
         conduit.pull(limit=self.options.limit,
                      quiet=self.options.quiet,
                      verbose=self.options.verbose,
+                     flush=self.options.no_skip_latest,
                      authors=self.options.authors,
                      stop_on_unknown_author=\
                          self.options.stop_on_unknown_author)